Removed lots of warnings with max compiler warning level of Eclipse
Roughly 1300 warnings were removed overall. The removal was nearly pure
refactoring (e.g., referencing this, removing redundant generic
parameters). In very few cases (~20) we improved the resource handling
using the try-with-resource construct that was added to Java in version
7. However, this also does not change any logic, but only improves the
resource handling in cases of failures to ensure that all resources are
closed properly.