Added dist-untestedjaronly as target to built script.
The new target is for only creating a new jar and updating the libraries
in the dist folder. This can save the time of executing the tests and
generating the Javadoc. Should be used carefully and only if it is made
sure that the tests are running and the up-to-date Javadoc is not
required.