New option to load bug data as bug counts instead of binary labels.
This new option impacts all data loaders. In the XML-definition, the new
attribute "classtype" can now be either "binary" or "numeric". In case
of binary, everything stays as it was and the data is loaded as nominal
true/false labels. In case of "numeric", bugs counts are loaded, so far
as they are available in the data. In case the original data only has
nominal labels, a warning is displayed and the numeric data will just
have the values 0 and 1. 

In case the "classtype" attribute is not specified, the default is
"binary". This ensure downwards compatibility with already existing
configuration files.