Added corner-case handling to LACE2
It was possible that the resampling to 100 instances for the nearest
unlike neighbor selection only drew instances from one class. This would
mean that there are no nearest unlike neighbors for the other class,
breaking the subsequent distance calculation. We changed the algorithm
to automatically resample in case only instances from one class are
drawn.