G-Measure now evaluates to 0.0 if recall=0 and pf=1
Before, this led to a NaN due to a division by zero. 0.0 makes sense, as
zero and pf of 1.0 both mean that the nominator will also be 0.0, i.e.,
we have 0.0/0.0. Moreover, this makes sense conceptually, because the
performance is as bad as possible in terms of what G-measure should
measure.